\name{BMLmodel-package}
\alias{BMLmodel-package}
\alias{BMLmodel}
\docType{package}
\title{
Simulating Biham-Middleton-Levine Traffic Models
}
\description{
This package is used to simulate simple traffic flow models (BML models) by specifying grid sizes, the number of red and blue cars and the number of steps. 
This package includes 6 functions: CreateGrid, RunBMLgrid, summary.BMLgrid, plot.BMLgrid, MoveCars and location_car.
}
\details{
\tabular{ll}{
Package: \tab BMLmodel\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-05-01\cr
License: \tab GPL-2\cr
}
Based on predefined rules by BLM models, a certain number of blue and cars move on a r*c dimensional grid. Traffic behavior will be simulated when total number of blue and red cars, steps and grid dimension are specified by callers.
}
\author{
Hong Fan

Maintainer: Hong Fan <honfan@ucdavis.edu>

}
\references{
T. Nagatani. The physics of traffic jams. Rep. Prog. Phys., 65(9):1331¨C1386, 2002.

R. M. D¡¯Souza. Coexisting phases and lattice dependence of a cellular automata model for traffic flow. Phys. Rev. E, 71, 2005.
}
\keyword{ package }

